refactor(web): make the loss-acknowledgement model surface-agnostic (MTN-150)
MTN-199 built the acknowledge-before-sign machinery inside `components/bridge/`,
with the transfer's identity spread across six typed fields — provider, both
chains, both asset addresses, the input amount. The trade surfaces need the same
frozen-basis model but have none of those fields, so the identity has to stop
being bridge-shaped before the gate can be reused.

Move the module to `components/loss-acknowledgement/` and collapse those six
fields into one opaque `identityKey` string that each surface composes for
itself; the six-way compare becomes a single string compare and the
`@osmosis-labs/bridge` import disappears. `swapImpactUnknown` becomes optional,
since only a bridge quote can bundle a swap of unknown impact.

`deriveMemoFlags` is renamed `deriveBridgeMemoFlags`. Its mapping is
surface-specific rather than shared: `LossFigures.slippage` is a realized total
loss on the bridge and stamps `loss=`, but the same field is a tolerance on a
trade and must stamp `slip=`. Leaving it unqualified in a now-shared module
invited exactly the conflation the separate memo keys exist to prevent.

No behaviour changes. The identity assertions necessarily changed shape — six
per-field cases became one key-compare case — but every other assertion in the
three moved specs passes untouched, and the bridge's own review-screen gate spec
passes unchanged.
